requirements.txt
# devops_jestparser_result.py file dependencies
beautifulsoup4==4.8.1
Click==7.0
elasticsearch==7.1.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
requests==2.23.0
gitpython==3.1.2
